pyproject.toml
[tool.poetry]
name = "impa-cast"
version = "0.1.0"
description = ""
authors = ["acataos <acatao1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.0.0"
einops = "^0.8.0"
timm = "^1.0.11"
wandb = "^0.18.6"
boto3 = "^1.35.56"
fvcore = "^0.1.5.post20221221"
torchmetrics = "^1.5.2"
scipy = "^1.14.1"
nvitop = "^1.3.2"
pytorch-lightning = "2.0"
joblib = "^1.4.2"
scikit-image = "^0.24.0"
moviepy = "^1.0.3"
pre-commit = "^2.20.0"
tensorboard = "^2.18.0"
pysteps = "^1.13.0"
h5py = "^3.12.1"
opencv-python = "4.8.1.78"
beartype = "^0.19.0"
numpy = "<2.0.0"
pandas = "^2.2.3"
plotly = "^5.24.1"
kaleido = "0.2.1"
xarray = "^2025.3.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
cartopy = "^0.24.1"
netcdf4 = "^1.7.2"
xr = "^1.0.5"
matplotlib = "3.9.2"
pyarrow = "^20.0.0"
jupyter = "^1.1.1"
diffusers = "^0.34.0"
satpy = "^0.57.0"


[tool.poetry.dev-dependencies]
pynvim = "^0.4.3"
pre-commit = "^2.20.0"
black = "^22.6.0"
isort = "^5.10.1"
pyproject-flake8 = "^0.0.1-alpha.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
profile = "black"

[tool.flake8]
max-line-length = 120
max-complexity = 20
exclude = ".git"
select = "B,C,E,F,W,T4,B9"
ignore = "E203, E266, E501, W503, F403, F401"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
